Export a command to address the color lose problem.
Some plugins may cause the color lose problem, like minibufexpl.

let g:miniBufExplForceSyntaxEnable = 1
This setting is the main cause. But it is a neccessary option to
address the highlight problem of minibufexpl.

I export a command from powerline, and call this command in
minibufexpl. The color lose problem is addressed perfectly.